refactor(alpha-checkbox): move directory
